--- src/models/settingType.model.ts
@@ -1,6 +1,0 @@
-export enum UISettingType {
-  TextInput = "textinput",
-  NumberInput = "numberinput",
-  Checkbox = "checkbox",
-  TextArea = "textarea"
-}
