+++ src/service/webview.service.ts
@@ -2,9 +2,9 @@ import { readFileSync } from "fs-extra";
 import { has, set } from "lodash";
 import { URL } from "url";
 import * as vscode from "vscode";
-import { CustomConfig } from "../models/customConfig.model";
-import { ExtensionConfig } from "../models/extensionConfig.model";
-import { UISettingType } from "../models/settingType.model";
+import { CustomConfig } from "../models/custom-config.model";
+import { ExtensionConfig } from "../models/extension-config.model";
+import { UISettingType } from "../models/setting-type.model";
 import { IWebviewSetting } from "../models/webview-setting.model";
 import { IWebview } from "../models/webview.model";
 import { state } from "../state";
